frontend/requirements.txt
flask
requests
python-dotenv
python-jose[cryptography]